# internal/api/docs/swagger.yaml
basePath: /v1
definitions:
  admin_auth_handler.SignInRequestDto:
    properties:
      email:
        example: ticketsale@gmail.com
        type: string
      password:
        example: "123"
        type: string
    required:
    - email
    - password
    type: object
  admin_auth_handler.SignInResponseDto:
    properties:
      code:
        example: 200
        type: integer
      expire:
        example: "2024-06-30T20:46:13-03:00"
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MTk3OTExNzMsImlkIjozLCJvcmlnX2lhdCI6MTcxOTcwNDc3Mywicm9sZSI6InVzZXIifQ.c8HuyRAxgNDC4FavwQ_mv-qWOm4Ch6--1-kSQEmK4x0
        type: string
    type: object
  admin_event_handler.CategoryDto:
    properties:
      description:
        example: EVENT
        type: string
      id:
        example: 3
        type: integer
      name:
        example: event
        type: string
    type: object
  admin_event_handler.CreateRequestDto:
    properties:
      city:
        example: Orlando
        type: string
      endAt:
        example: "2024-01-10T00:00:00Z"
        type: string
      location:
        example: Disney
        type: string
      product:
        $ref: '#/definitions/admin_product_handler.CreateRequestDto'
      startAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      state:
        example: FL
        type: string
    type: object
  admin_event_handler.CreateResponseDto:
    properties:
      city:
        example: Orlando
        type: string
      endAt:
        example: "2024-01-10T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: Disney
        type: string
      product:
        $ref: '#/definitions/admin_product_handler.CreateResponseDto'
      productId:
        example: 1
        type: integer
      startAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      state:
        example: FL
        type: string
    type: object
  admin_event_handler.GetAllProductDto:
    properties:
      active:
        example: true
        type: boolean
      category:
        $ref: '#/definitions/admin_event_handler.CategoryDto'
      categoryId:
        example: 3
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        type: string
      discountPrice:
        example: 4.99
        type: number
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        type: string
      isDeleted:
        example: false
        type: boolean
      name:
        example: Red Hot Chilly Peppers
        type: string
      price:
        example: 5.99
        type: number
      uuid:
        example: 998f91f3-4dd7-419d-a543-0d26a0e945ec
        type: string
    type: object
  admin_event_handler.GetAllResponseDto:
    properties:
      city:
        example: Orlando
        type: string
      endAt:
        example: "2024-01-10T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: Disney
        type: string
      product:
        $ref: '#/definitions/admin_event_handler.GetAllProductDto'
      productId:
        example: 1
        type: integer
      startAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      state:
        example: FL
        type: string
    type: object
...
  admin_event_handler.GetOneByIdResponseDto:
    properties:
      city:
        example: Orlando
        type: string
      endAt:
        example: "2024-01-10T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: Disney
        type: string
      product:
        $ref: '#/definitions/admin_event_handler.GetOneByIdProductDto'
      productId:
        example: 1
        type: integer
      startAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      state:
        example: FL
        type: string
    type: object
...
  admin_event_handler.UpdateRequestDto:
    properties:
      city:
        example: Orlando
        type: string
      endAt:
        example: "2024-01-10T00:00:00Z"
        type: string
      location:
        example: Disney
        type: string
      product:
        $ref: '#/definitions/admin_product_handler.UpdateRequestDto'
      startAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      state:
        example: FL
        type: string
    type: object
  admin_gateway_handler.CreatePaymentTypeResDTO:
    properties:
      gatewayId:
        example: 1
        type: integer
      gatewayPaymentTypeId:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
    type: object
  admin_gateway_handler.CreateReqDTO:
    properties:
      active:
        example: true
        type: boolean
      adqCode3ds:
        example: "123456"
        type: string
      authType:
        enum:
        - bearer
        - basic
        example: bearer
        type: string
      clientId:
        example: ZYDPLLBWSK3MVQJSIYHB1OR2JXCY0X2C5UJ2QAR2MAAIT5Q
        type: string
      clientSecret:
        example: 52caa7c6-d107-42bd-afac-0226a501e66e
        type: string
      defaultAdqCode:
        example: "654321"
        type: string
      description:
        example: Payment gateway for Stripe
        type: string
      gatewayProcessId:
        example: 1
        minimum: 1
        type: integer
      gatewayProviderId:
        example: 1
        minimum: 1
        type: integer
      name:
        example: Stripe
        minLength: 3
        type: string
      notifPassword:
        example: notification_password
        type: string
      notifUser:
        example: notification_user
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
      paymentTypes:
        example:
        - 1
        - 2
        items:
          type: integer
        minItems: 1
        type: array
      softDescriptor:
        example: Company Name
        type: string
      testEnvironment:
        example: false
        type: boolean
      url:
        example: https://api.stripe.com
        type: string
      use3ds:
        example: false
        type: boolean
      useAntifraud:
        example: true
        type: boolean
      webhookUrl:
        example: https://example.com/webhook
        type: string
    required:
    - authType
    - gatewayProcessId
    - gatewayProviderId
    - name
    - order
    - paymentTypes
    type: object
  admin_gateway_handler.CreateResDTO:
    properties:
      active:
        example: true
        type: boolean
      adqCode3ds:
        example: "123456"
        type: string
      authType:
        example: bearer
        type: string
      clientId:
        example: ZYDPLLBWSK3MVQJSIYHB1OR2JXCY0X2C5UJ2QAR2MAAIT5Q
        type: string
      clientSecret:
        example: 52caa7c6-d107-42bd-afac-0226a501e66e
        type: string
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      defaultAdqCode:
        example: "654321"
        type: string
      description:
        example: Payment gateway for Stripe
        type: string
      gatewayProcessId:
        example: 1
        type: integer
      gatewayProviderId:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      isDeleted:
        example: false
        type: boolean
      name:
        example: Stripe
        type: string
      notifPassword:
        example: notification_password
        type: string
      notifUser:
        example: notification_user
        type: string
      order:
        example: 1
        type: integer
      paymentTypes:
        items:
          $ref: '#/definitions/admin_gateway_handler.CreatePaymentTypeResDTO'
        type: array
      softDescriptor:
        example: Company Name
        type: string
      testEnvironment:
        example: false
        type: boolean
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      url:
        example: https://api.stripe.com
        type: string
      use3ds:
        example: false
        type: boolean
      useAntifraud:
        example: true
        type: boolean
      uuid:
        example: 78e5259f-1b4e-460c-90d2-83e640a7d024
        type: string
      webhookUrl:
        example: https://example.com/webhook
        type: string
    type: object
  admin_gateway_handler.GatewayPaymentTypesDTO:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Installment
        type: string
    type: object
  admin_gateway_handler.GatewayProcessDTO:
    properties:
      id:
        example: 2
        type: integer
      name:
        example: Authorize Only
        type: string
    type: object
  admin_gateway_handler.GetAllResDTO:
    properties:
      active:
        example: true
        type: boolean
      adqCode3ds:
        example: "123456"
        type: string
      authType:
        example: bearer
        type: string
      clientId:
        example: ZYDPLLBWSK3MVQJSIYHB1OR2JXCY0X2C5UJ2QAR2MAAIT5Q
        type: string
      clientSecret:
        example: 52caa7c6-d107-42bd-afac-0226a501e66e
        type: string
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      defaultAdqCode:
        example: "654321"
        type: string
      description:
        example: Payment gateway for Stripe
        type: string
      gatewayPaymentTypes:
        items:
          $ref: '#/definitions/admin_gateway_handler.GatewayPaymentTypesDTO'
        type: array
      gatewayProcess:
        $ref: '#/definitions/admin_gateway_handler.GatewayProcessDTO'
      gatewayProcessId:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Stripe
        type: string
      notifPassword:
        example: notification_password
        type: string
      notifUser:
        example: notification_user
        type: string
      order:
        example: 1
        type: integer
      softDescriptor:
        example: Company Name
        type: string
      testEnvironment:
        example: false
        type: boolean
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      url:
        example: https://api.stripe.com
        type: string
      use3ds:
        example: false
        type: boolean
      useAntifraud:
        example: true
        type: boolean
      uuid:
        example: 78e5259f-1b4e-460c-90d2-83e640a7d024
        type: string
      webhookUrl:
        example: https://example.com/webhook
        type: string
    type: object
  admin_gateway_handler.GetOneByIdResDTO:
    properties:
      active:
        example: true
        type: boolean
      adqCode3ds:
        example: "123456"
        type: string
      authType:
        example: bearer
        type: string
      clientId:
        example: ZYDPLLBWSK3MVQJSIYHB1OR2JXCY0X2C5UJ2QAR2MAAIT5Q
        type: string
      clientSecret:
        example: 52caa7c6-d107-42bd-afac-0226a501e66e
        type: string
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      defaultAdqCode:
        example: "654321"
        type: string
      description:
        example: Payment gateway for Stripe
        type: string
      gatewayPaymentTypes:
        items:
          $ref: '#/definitions/admin_gateway_handler.GatewayPaymentTypesDTO'
        type: array
      gatewayProcess:
        $ref: '#/definitions/admin_gateway_handler.GatewayProcessDTO'
      gatewayProcessId:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Stripe
        type: string
      notifPassword:
        example: notification_password
        type: string
      notifUser:
        example: notification_user
        type: string
      order:
        example: 1
        type: integer
      softDescriptor:
        example: Company Name
        type: string
      testEnvironment:
        example: false
        type: boolean
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      url:
        example: https://api.stripe.com
        type: string
      use3ds:
        example: false
        type: boolean
      useAntifraud:
        example: true
        type: boolean
      uuid:
        example: 78e5259f-1b4e-460c-90d2-83e640a7d024
        type: string
      webhookUrl:
        example: https://example.com/webhook
        type: string
    type: object
  admin_gateway_handler.UpdateReqDTO:
    properties:
      active:
        example: true
        type: boolean
      adqCode3ds:
        example: "123456"
        type: string
      authType:
        enum:
        - bearer
        - basic
        example: bearer
        type: string
      clientId:
        example: ZYDPLLBWSK3MVQJSIYHB1OR2JXCY0X2C5UJ2QAR2MAAIT5Q
        type: string
      clientSecret:
        example: 52caa7c6-d107-42bd-afac-0226a501e66e
        type: string
      defaultAdqCode:
        example: "654321"
        type: string
      description:
        example: Payment gateway for Stripe
        type: string
      gatewayProcessId:
        example: 1
        minimum: 1
        type: integer
      gatewayProviderId:
        example: 1
        minimum: 1
        type: integer
      name:
        example: Stripe
        minLength: 3
        type: string
      notifPassword:
        example: notification_password
        type: string
      notifUser:
        example: notification_user
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
      softDescriptor:
        example: Company Name
        type: string
      testEnvironment:
        example: false
        type: boolean
      url:
        example: https://api.stripe.com
        type: string
      use3ds:
        example: false
        type: boolean
      useAntifraud:
        example: true
        type: boolean
      webhookUrl:
        example: https://example.com/webhook
        type: string
    required:
    - authType
    - gatewayProcessId
    - gatewayProviderId
    - name
    - order
    type: object
...
  exception.HttpException:
    properties:
      message:
        example: Invalid path parameter
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
  middleware.AuthenticationError:
    properties:
      message:
        example: Access denied
        type: string
      statusCode:
        example: 401
        type: integer
    type: object
  middleware.RolePermissionError:
    properties:
      message:
        example: permission denied.
        type: string
      statusCode:
        example: 403
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Ticket Sale Support
    url: http://www.swagger.io/support
  description: This is a simple ticket sales application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ticket Sale
  version: "1.0"
paths:
  /admin/auth:
    post:
      consumes:
      - application/json
      description: Sign In
      parameters:
      - description: Sign In
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/admin_auth_handler.SignInRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_auth_handler.SignInResponseDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
      summary: Sign In
      tags:
      - Admin Users
  /admin/events:
    get:
      description: Get all events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_event_handler.GetAllResponseDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get all events
      tags:
      - Admin Event
    post:
      description: Create a event
      parameters:
      - description: New Event
        in: body
        name: Event
        required: true
        schema:
          $ref: '#/definitions/admin_event_handler.CreateRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin_event_handler.CreateResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Create a event
      tags:
      - Admin Event
  /admin/events/{id}:
    delete:
      description: Soft Delete a event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Soft Delete a event
      tags:
      - Admin Event
    get:
      description: Get One By Id
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_event_handler.GetOneByIdResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get One By Id
      tags:
      - Admin Event
    put:
      description: Update a event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Event
        in: body
        name: Event
        required: true
        schema:
          $ref: '#/definitions/admin_event_handler.UpdateRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Update a event
      tags:
      - Admin Event
  /admin/gateway:
    get:
      description: Get all gateway
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_gateway_handler.GetAllResDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get all gateway
      tags:
      - Admin Gateway
    post:
      description: Create a gateway
      parameters:
      - description: New Gateway
        in: body
        name: Gateway
        required: true
        schema:
          $ref: '#/definitions/admin_gateway_handler.CreateReqDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin_gateway_handler.CreateResDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Create a gateway
      tags:
      - Admin Gateway
  /admin/gateway/{id}:
    delete:
      description: Soft Delete a gateway
      parameters:
      - description: Gateway ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Soft Delete a gateway
      tags:
      - Admin Gateway
    get:
      description: Get One By Id
      parameters:
      - description: Gateway ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_gateway_handler.GetOneByIdResDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get One By Id
      tags:
      - Admin Gateway
    put:
      description: Update a gateway
      parameters:
      - description: Gateway ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update gateway
        in: body
        name: Gateway
        required: true
        schema:
          $ref: '#/definitions/admin_gateway_handler.UpdateReqDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Update a gateway
      tags:
      - Admin Gateway
...
swagger: "2.0"